+++ src/main/steps/common/task-list/components/tasklist/config/index.ts
@@ -1,0 +1,3 @@
+import { CA_APPLICANT } from './ca_applicant';
+import { DA_APPLICANT } from './da_applicant';
+export default { CA_APPLICANT, DA_APPLICANT };
